Refactor: fix packages location for a couple of unit tests
